Added better console logging and minor temp changes.

+++ db_model.js
@@ -1,9 +1,9 @@
 const Database = require('better-sqlite3')
 
 function createDatabase(guildId) {
-    // console.log(`createDatabase(${guildId}) called.`)
+    console.log(`createDatabase(${guildId}) called.`)
 
-    let db = new Database(`./databases/${guildId}.sqlite`);
+    let db = new Database(`./databases/${guildId}.sqlite`, {verbose: console.log});
 
     const createStatements = [
         `CREATE TABLE IF NOT EXISTS emojiActivity(emoji TEXT, person TEXT, datetime TEXT)`,
@@ -26,44 +26,52 @@ function createDatabase(guildId) {
     db.close()
 }
 
-function deleteFromDb(guildId, emojiId, personId, dateTime) {
-    // console.log(`deleteFromDb(${guildId}, ${emojiId}) called.`)
-    let db = new Database(`./databases/${guildId}.sqlite`);
-    const statement = db.prepare(`
-        DELETE FROM emojiActivity 
-        WHERE rowid = 
-        (
-            SELECT rowid
-            FROM emojiActivity
-            WHERE
-                emoji = @emoji AND
-                person = @person AND
-                datetime = @datetime
-            LIMIT 1
-        )
-    `)
-    statement.run({
-        emoji: emojiId,
-        person: personId,
-        datetime: dateTime
-    })
-    db.close()
+function deleteFromDb(guildId, emojiId, personId, dateTime, origin) {
+    console.log(`deleteFromDb(${guildId}, ${emojiId}, ${personId}, ${dateTime}) called from ${origin}.`)
+    if (guildId && emojiId && personId && dateTime) {
+        let db = new Database(`./databases/${guildId}.sqlite`);
+        const statement = db.prepare(`
+            DELETE FROM emojiActivity 
+            WHERE rowid = 
+            (
+                SELECT rowid
+                FROM emojiActivity
+                WHERE
+                    emoji = @emoji AND
+                    person = @person AND
+                    datetime = @datetime
+                LIMIT 1
+            )
+        `)
+        statement.run({
+            emoji: emojiId,
+            person: personId,
+            datetime: dateTime
+        })
+        db.close()
+    } else {
+        console.log("deleteFromDb failed.")
+    }
 
 }
 
-function insertToDb(guildId, emojiId, personId, dateTime) {
-    // console.log(`insertToDb(${guildId}, ${emojiId}) called.`)
-    let db = new Database(`./databases/${guildId}.sqlite`);
-    const statement = db.prepare(`
-        INSERT INTO emojiActivity (emoji, person, datetime) 
-        VALUES (@emoji, @person, @datetime)
-    `)
-    statement.run({
-        emoji: emojiId,
-        person: personId,
-        datetime: dateTime
-    })
-    db.close()
+function insertToDb(guildId, emojiId, personId, dateTime, origin) {
+    console.log(`insertToDb(${guildId}, ${emojiId}, ${personId}, ${dateTime}) called from ${origin}.`)
+    if (guildId && emojiId && personId && dateTime) {
+        let db = new Database(`./databases/${guildId}.sqlite`);
+        const statement = db.prepare(`
+            INSERT INTO emojiActivity (emoji, person, datetime) 
+            VALUES (@emoji, @person, @datetime)
+        `)
+        statement.run({
+            emoji: emojiId,
+            person: personId,
+            datetime: dateTime
+        })
+        db.close()
+    } else {
+        console.log("insertToDb failed.")
+    }
 
 }
